Refactor page details to support `react-router` v6
