+++ app/routes/admin/groups/components/GroupForm/index.tsx
@@ -44,7 +44,6 @@ type Props = {
 
 const GroupForm = ({ isInterestGroup }: Props) => {
   const { groupId } = useParams<{ groupId: string }>();
-  console.log('groupId', groupId);
   const group = useAppSelector((state) => selectGroup(state, { groupId }));
   const isNew = !group;
